More tests. Constants created from doubles nor ints

+++ logic-language/framework/src/main/java/com/dan323/expresions/base/Constant.java
@@ -5,5 +5,5 @@ package com.dan323.expresions.base;
  */
 public interface Constant extends LogicOperation {
 
-    Constant construct(int val);
+    Constant construct(double val);
 }
